pyproject.toml
[tool.poetry]
name = "social-network"
version = "0.1.0"
description = "A team educational project to create a social network."
authors = ["Дмитрий Афанасьев <wackymc@yandex.ru>", "Никита Иванович nik.popkov.98@list.ru"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.12" }
uvicorn = "^0.34.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
pydantic-settings = "^2.8.1"
alembic = "^1.15.1"
asyncpg = "^0.30.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.40" }
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
jinja2 = "^3.1.6"
redis = "^5.2.1"
celery = {extras = ["redis"], version = "^5.5.2"}
aio-pika = "^9.5.5"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
pytest-asyncio = "^1.0.0"
pytest = "^8.4.0"
pytest-dotenv = "^0.5.2"
asgi-lifespan = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

